Separating functions from the main script
